requirements.txt
moviepy==2.2.1
numpy==2.3.1
Pillow==11.3.0
